# refactoring: replace a byte.Buffer with a byte slice