This change also removes `length` member with the len of the slice.

This slightly make the audio decoding performance worse, but I am not sure why.

```
goos: darwin
goarch: arm64
pkg: github.com/gen2brain/mpeg
               │   old.txt   │              new.txt               │
               │   sec/op    │   sec/op     vs base               │
DecodeVideo-12   76.90µ ± 1%   76.34µ ± 1%  -0.73% (p=0.024 n=10)
DecodeAudio-12   47.64µ ± 1%   49.36µ ± 2%  +3.62% (p=0.002 n=10)
RGBA-12          41.29µ ± 0%   40.53µ ± 0%  -1.83% (p=0.001 n=10)
geomean          53.28µ        53.45µ       +0.32%

               │  old.txt   │              new.txt              │
               │    B/op    │    B/op     vs base               │
DecodeVideo-12   76.00 ± 1%   75.00 ± 0%  -1.32% (p=0.008 n=10)
DecodeAudio-12   46.50 ± 1%   48.00 ± 2%  +3.23% (p=0.000 n=10)
RGBA-12          40.00 ± 2%   40.00 ± 0%       ~ (p=0.582 n=10)
geomean          52.09        52.41       +0.62%

               │   old.txt    │               new.txt               │
               │  allocs/op   │ allocs/op   vs base                 │
DecodeVideo-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecodeAudio-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
RGBA-12          0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                     ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```