@gen2brain 2024-04-12
I remember I switched back and forth between byte slice and bytes.Buffer, I can't remember what was the difference and why bytes.Buffer is finally used.