@hajimehoshi 2024-04-12
From my intuition, the current byte.Buffer just adds overheads, but I might be wrong.